[IN-PROGRESS,NON-COMPILIING]: Starting to remove the unnecessary fields from code for using async storage
