+++ src-tauri/src/data_controller.rs
@@ -4,6 +4,7 @@ use openhome_core::{Error, Result};
 use std::path::PathBuf;
 use tauri::AppHandle;
 
+
 pub struct TauriDataController(tauri::AppHandle);
 impl TauriDataController {
     pub fn new(handle: &tauri::AppHandle) -> Self {
