+++ src-tauri/src/lib.rs
@@ -7,25 +7,25 @@ mod plugin;
 mod startup;
 mod startup_config;
 mod state;
-mod synced_state;
+mod async_state;
 mod util;
 mod version;
 
 use crate::data_controller::ToDataController;
-use crate::synced_state::AllSyncedState;
+use crate::async_state::LazyState;
 use openhome_core::convert_strategies::ConvertStrategies;
 use openhome_core::lookup::LookupState;
-use openhome_core::ohpkm_store::OhpkmBytesStore;
 use openhome_core::{Error, Result};
-use pkm_rs::ohpkm::OhpkmV2;
 use std::env;
 use tauri::Manager;
+use box_pointer::BoxPointer;
+use openhome_core::pkm_storage::{load_banks, StoredBankData};
 
 const RAW_HANDLER: fn(tauri::ipc::Invoke<tauri::Wry>) -> bool = tauri::generate_handler![
     commands::get_file_bytes,
     commands::get_storage_file_json,
     commands::write_storage_file_json,
-    synced_state::ohpkm_store::add_to_ohpkm_store,
+    async_state::ohpkm_store::add_to_ohpkm_store,
     logging::log,
 ];
 
@@ -54,15 +54,15 @@ pub fn specta_builder() -> tauri_specta::Builder<tauri::Wry> {
         state::start_transaction,
         state::rollback_transaction,
         state::commit_transaction,
-        synced_state::save_synced_state,
-        synced_state::update_synced_state,
-        synced_state::convert_strategies::get_convert_strategies,
-        synced_state::convert_strategies::update_convert_strategies,
-        synced_state::lookup::get_lookups,
-        synced_state::lookup::add_to_lookups,
-        synced_state::lookup::remove_dangling,
-        synced_state::ohpkm_store::get_ohpkm_store,
-        synced_state::ohpkm_store::permanently_delete_ohpkms,
+        async_state::save_synced_state,
+        async_state::update_synced_state,
+        async_state::convert_strategies::get_convert_strategies,
+        async_state::convert_strategies::update_convert_strategies,
+        async_state::lookup::get_lookups,
+        async_state::lookup::add_to_lookups,
+        async_state::lookup::remove_dangling,
+        async_state::ohpkm_store::get_ohpkm_store,
+        async_state::ohpkm_store::permanently_delete_ohpkms,
         logging::get_logs_today,
         logging::clear_logs_for_range,
     ])
@@ -106,10 +106,11 @@ pub fn run() {
 
             let controller = app.handle().controller();
 
-            let ohpkm_store = match OhpkmBytesStore::load_from_mons_v2(&controller) {
-                Ok(state) => state,
+            // TODO: Only get information on the current bank and load that one.
+            let mut bank_info = match load_banks(&controller) {
+                Ok(bank_info) => bank_info,
                 Err(err) => {
-                    util::show_error_dialog(app, err, launch_error_msg("OHPKM Load"));
+                    util::show_error_dialog(app, err, launch_error_msg("Unable to obtain information on your Pokemon bank info!"));
 
                     app.handle().exit(1);
                     std::process::exit(1);
@@ -126,12 +127,6 @@ pub fn run() {
                 }
             };
 
-            lookup_state.with_recalculated(
-                ohpkm_store
-                    .all_entries()
-                    .filter_map(|(_, bytes)| OhpkmV2::from_bytes(bytes).ok()),
-            );
-
             let conversion_settings = match ConvertStrategies::load_from_storage(&controller) {
                 Ok(settings) => settings,
                 Err(err) => {
@@ -143,7 +138,7 @@ pub fn run() {
             };
 
             let synced_state =
-                AllSyncedState::from_states(lookup_state, ohpkm_store, conversion_settings);
+                LazyState::from_partial_states(lookup_state, get_bank_pointer(bank_info), conversion_settings);
             app.manage(synced_state);
 
             let pokedex_state = match state::PokedexState::load_from_storage(&controller) {
@@ -186,6 +181,12 @@ pub fn run() {
         })
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
+}
+
+fn get_bank_pointer(bank_data: StoredBankData) -> BoxPointer {
+
+    return bank_data.get_current_box_pointer();
+
 }
 
 fn launch_error_msg(error_category: &str) -> String {
